Fix last few Py_ssize_t issues.
https://sourceforge.net/tracker/index.php?func=detail&aid=1680699&group_id=22307&atid=374934

+++ _mysql.c
@@ -223,7 +223,7 @@ static PyObject *_mysql_server_init(
 	PyObject *kwargs) {
 	static char *kwlist[] = {"args", "groups", NULL};
 	char **cmd_args_c=NULL, **groups_c=NULL, *s;
-	int cmd_argc=0, i, groupc;
+	Py_ssize_t cmd_argc=0, i, groupc;
 	PyObject *cmd_args=NULL, *groups=NULL, *ret=NULL, *item;
 
 	if (_mysql_server_init_done) {
@@ -395,7 +395,7 @@ _mysql_ResultObject_Initialize(
 			Py_INCREF(Py_None);
 		}
 		if (PySequence_Check(fun)) {
-			int j, n2=PySequence_Size(fun);
+			Py_ssize_t j, n2=PySequence_Size(fun);
 			PyObject *fun2=NULL;
 			for (j=0; j<n2; j++) {
 				PyObject *t = PySequence_GetItem(fun, j);
